chore(core): backend restructuration: telemetry (#4679)
