+++ src/bp/core/botpress.ts
@@ -7,6 +7,7 @@ import { SessionIdFactory } from 'core/dialog/sessions'
 import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine } from 'core/events'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
 import { copyDir } from 'core/misc/pkg-fs'
+import { StatsService } from 'core/telemetry'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
 import { inject, injectable, tagged } from 'inversify'
@@ -47,7 +48,6 @@ import { NotificationsService } from './services/notification/service'
 import RealtimeService from './services/realtime'
 import { DataRetentionJanitor } from './services/retention/janitor'
 import { DataRetentionService } from './services/retention/service'
-import { StatsService } from './services/stats-service'
 import { WorkspaceService } from './services/workspace-service'
 import { Statistics } from './stats'
 import { TYPES } from './types'
