+++ src/bp/core/repositories/repositories.inversify.ts
@@ -2,17 +2,14 @@ import { StrategyUsersRepository, WorkspaceInviteCodesRepository, WorkspaceUsers
 import { SessionRepository } from 'core/dialog/sessions'
 import { EventRepository } from 'core/events'
 import { LogsRepository } from 'core/logger'
+import { TelemetryRepository } from 'core/telemetry'
 import { ContainerModule, interfaces } from 'inversify'
 
 import { TYPES } from '../types'
-
 import { KnexNotificationsRepository, KnexUserRepository, NotificationsRepository, UserRepository } from '.'
-
 import { ConversationRepository, KnexConversationRepository } from './conversations'
-
 import { KnexMessageRepository, MessageRepository } from './messages'
 import { TasksRepository } from './tasks'
-import { TelemetryRepository } from './telemetry'
 
 const RepositoriesContainerModule = new ContainerModule((bind: interfaces.Bind) => {
   bind<SessionRepository>(TYPES.SessionRepository)
