+++ src/bp/core/server.ts
@@ -10,6 +10,7 @@ import { CMSService, CMSRouter } from 'core/cms'
 import { ExternalAuthConfig, ConfigProvider } from 'core/config'
 import { ConverseRouter, ConverseService } from 'core/converse'
 import { LogsService, LogsRepository } from 'core/logger'
+import { TelemetryRouter, TelemetryRepository } from 'core/telemetry'
 import cors from 'cors'
 import errorHandler from 'errorhandler'
 import { UnlicensedError } from 'errors'
@@ -32,7 +33,6 @@ import { URL } from 'url'
 import yn from 'yn'
 
 import { ModuleLoader } from './module-loader'
-import { TelemetryRepository } from './repositories/telemetry'
 import { AdminRouter, AuthRouter, BotsRouter, MediaRouter, ModulesRouter } from './routers'
 import { HintsRouter } from './routers/bots/hints'
 import { NLURouter } from './routers/bots/nlu'
@@ -40,7 +40,6 @@ import { isDisabled } from './routers/conditionalMiddleware'
 import { InvalidExternalToken, PaymentRequiredError } from './routers/errors'
 import { SdkApiRouter } from './routers/sdk/router'
 import { ShortLinksRouter } from './routers/shortlinks'
-import { TelemetryRouter } from './routers/telemetry'
 import { hasPermissions, monitoringMiddleware, needPermissions } from './routers/util'
 import { GhostService } from './services'
 import ActionServersService from './services/action/action-servers-service'
