+++ src/bp/core/telemetry/index.ts
@@ -1,0 +1,4 @@
+export * from './stats-service'
+export * from './telemetry-repository'
+export * from './telemetry-router'
+export * from './stats/telemetry.inversify'
