+++ src/bp/core/telemetry/stats/hooks-stats.ts
@@ -3,16 +3,15 @@ import LicensingService from 'common/licensing-service'
 import { buildSchema } from 'common/telemetry'
 import Database from 'core/database'
 import { calculateHash } from 'core/misc/utils'
-import { TelemetryRepository } from 'core/repositories/telemetry'
+import { GhostService } from 'core/services'
+import { BotService } from 'core/services/bot-service'
+import { JobService } from 'core/services/job-service'
 import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import _ from 'lodash'
 import ms from 'ms'
 
-import { GhostService } from '..'
-import { BotService } from '../bot-service'
-import { JobService } from '../job-service'
-
+import { TelemetryRepository } from '../telemetry-repository'
 import { TelemetryStats } from './telemetry-stats'
 
 export interface BotHooks {
@@ -71,7 +70,7 @@ const parseHookName = (hookName: string, isBuiltIn: boolean) => {
 }
 
 @injectable()
-export class HooksLifecycleStats extends TelemetryStats {
+export class HooksStats extends TelemetryStats {
   protected interval: number
   protected url: string
   protected lock: string
