+++ src/bp/core/telemetry/stats/roles-stats.ts
@@ -4,15 +4,14 @@ import { buildSchema } from 'common/telemetry'
 import { AuthRule, Workspace } from 'common/typings'
 import Database from 'core/database'
 import { calculateHash } from 'core/misc/utils'
-import { TelemetryRepository } from 'core/repositories/telemetry'
+import { GhostService } from 'core/services'
+import { JobService } from 'core/services/job-service'
 import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import _ from 'lodash'
 import ms from 'ms'
 
-import { GhostService } from '..'
-import { JobService } from '../job-service'
-
+import { TelemetryRepository } from '../telemetry-repository'
 import { REDACTED, TelemetryStats } from './telemetry-stats'
 
 const DEFAULT_ROLES = ['dev', 'admin', 'editor', 'agent']
